+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Details - Display normals as deviation directly under daytime/nighttime temperature
 - Improve display of precipitation details
 - Details - Make tooltips persistent until you click outside the bounds of the tooltip
+- Details - Show current air quality on Today page when no daily air quality is available
 
 
 # Version 6.0.2-alpha (2025-07-19)